Fix the daily purge, which has never deleted a row
Spring Data rejects a long return type on a @Modifying query at invocation
time, so deleteByAssignedAtBefore threw every midnight:

  Modifying queries can only use void or int/Integer as return type;
  Offending method: ...AssignmentLogRepository.deleteByAssignedAtBefore

Four repositories had the same long return, and all four are called from
ShiftAssignService.purgeOldData(). Because the first call threw, the roster,
cover-routing and roster-event purges never ran either — the 7-day retention
was silently a no-op on every table, not just the activity log.

Changes long -> int on all four, and wraps each step so a future failure in
one table cannot stop the rest; a failed step logs and reports -1 instead of
aborting the run. The purge cron also becomes overridable
(jira.purge.cron, default unchanged at midnight IST) like the other two
schedulers, so a run can be triggered on demand rather than waiting a day.

Verified against H2: seeded three 30-day-old assignment_log rows, ran the
purge, log reported "activity log: 3 rows" and a direct query confirmed 0
remaining. Before the fix the same run logged only the exception above.

+++ src/main/java/com/jira/autoassign/repository/AssignmentLogRepository.java
@@ -21,8 +21,10 @@ public interface AssignmentLogRepository extends JpaRepository<AssignmentLog, Lo
     // "Assignment History" view. Filtered to ASSIGN so the cycle reads cleanly.
     List<AssignmentLog> findTop24ByTeamIdAndActionOrderByAssignedAtDescIdDesc(String teamId, String action);
 
+    // Must return void or int/Integer — Spring Data rejects a long return on a
+    // @Modifying query at invocation time, which silently killed the daily purge.
     @Modifying
     @Transactional
     @Query("DELETE FROM AssignmentLog l WHERE l.assignedAt < :cutoff")
-    long deleteByAssignedAtBefore(@Param("cutoff") LocalDateTime cutoff);
+    int deleteByAssignedAtBefore(@Param("cutoff") LocalDateTime cutoff);
 }
